Rename: raptor → vulture
The crate has practically nothing in common with raptor-rs 0.2 — substantially
rewritten across the v0.3-v0.14 arc and the rRAPTOR + restructure work that
followed. Forking under a new name to reflect that.

- Workspace member crates renamed:
  - raptor → vulture (main library)
  - raptor-proptest → vulture-proptest
  - raptor-dotgraph → vulture-dotgraph
- Workspace member directories renamed: raptor/ → vulture/, raptor-proptest/ → vulture-proptest/
- Repository URL updated to https://github.com/urschrei/vulture
- All raptor:: → vulture:: references in source, doctests, doc-comments, examples, benches
- README + roadmap + cross-city-benchmarks + vulture-proptest README updated
- soundness.md, docs/proptest.md, docs/superpowers/specs/, docs/superpowers/plans/, CHANGELOG.md left verbatim — historical record of when the crate was called raptor

The bench file vulture/benches/raptor.rs is unchanged: it benchmarks the
RAPTOR algorithm, named after what it tests, not after the crate.
The algorithm itself is and stays RAPTOR (the paper's name).
