+++ dotgraph/Cargo.toml
@@ -1,8 +1,8 @@
 [package]
-name = "raptor-dotgraph"
+name = "vulture-dotgraph"
 version = "0.1.0"
 edition = "2024"
 
 [dependencies]
 clap = { version = "4", features = ["derive"] }
-raptor = { workspace = true, features = ["dotgraph", "internal"] }
+vulture = { workspace = true, features = ["dotgraph", "internal"] }
