+++ dotgraph/src/main.rs
@@ -1,5 +1,5 @@
 use clap::{Parser, Subcommand};
-use raptor::manual::builders;
+use vulture::manual::builders;
 
 #[derive(Parser)]
 #[command(
