+++ vulture/examples/check-duplicates.rs
