+++ vulture/examples/check-monotonic.rs
