+++ vulture/examples/reboarding.rs
@@ -1,8 +1,8 @@
 //! Example showing a multi-route RAPTOR query where a passenger reboards
 //! a shared route at a later stop reached via a faster feeder route.
 
-use raptor::Duration;
-use raptor::{RouteIdx, SecondOfDay, StopIdx, Timetable, TripIdx};
+use vulture::Duration;
+use vulture::{RouteIdx, SecondOfDay, StopIdx, Timetable, TripIdx};
 
 // Stop indices: S=0, A=1, B=2, C=3, D=4
 // Route indices: R1=0 (S->A), R2=1 (S->B), R3=2 (A->B->C->D)
