+++ vulture/src/algorithm/boarding.rs
