+++ vulture/src/algorithm/footpaths.rs
