+++ vulture/src/algorithm/label_bag.rs
