+++ vulture/src/algorithm/mod.rs
