+++ vulture/src/algorithm/per_call.rs
