+++ vulture/src/algorithm/range.rs
