+++ vulture/src/journey.rs
