+++ vulture/src/label.rs
