+++ vulture/src/manual/mod.rs
