+++ vulture/src/query.rs
